+++ src/page/Category/SubCategory.jsx
@@ -8,6 +8,7 @@ import {
 } from "../../redux/features/category/categoryApi";
 import { useParams } from "react-router-dom";
 import { MdOutlineDeleteForever } from "react-icons/md";
+import Swal from "sweetalert2";
 
 const SubCategory = ({ id, subCategoryName }) => {
 
@@ -88,11 +89,31 @@ const SubCategory = ({ id, subCategoryName }) => {
 
     const handleDeleteCategory = async (category) => {
         try {
-            const res = await deleteCategory(category._id).unwrap();
-            if (res?.statusCode === 200) {
-                message.success(res.message);
-                refetch();
-            }
+
+            Swal.fire({
+                title: "Are you sure?",
+                text: "You won't be able to revert this!",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Yes, delete it!"
+            }).then(async (result) => {
+                if (result.isConfirmed) {
+                    const res = await deleteCategory(category._id).unwrap();
+                    if (res?.statusCode === 200) {
+                        message.success(res.message);
+                        refetch();
+                    }
+                    Swal.fire({
+                        title: "Deleted!",
+                        text: "Your file has been deleted.",
+                        icon: "success"
+                    });
+                }
+            });
+
+
         } catch (error) {
             console.error(error);
             message.error("Something went wrong");
